Big update with authorization and authentication

+++ controllers/avocatController.js
@@ -3,8 +3,18 @@ const Avocat = db.Avocat;
 const Domaine = db.Domaine;
 const Domaine_Avocat = db.DomaineAvocat;
 
-exports.avocatsListe = async function (req, res) {
-    await Avocat.findAll()
+
+
+
+exports.lawyersList = async function (req, res) {
+    await Avocat.findAll({
+        attributes: ['id_avocat', 'nom_avocat', 'coordonnees', 'honoraires', 'photo'],
+        include: [{
+            model: Domaine,
+            attributes: ['id_domaine', 'nom_domaine'],
+            through: { attributes: [] }
+        }]
+    })
     .then(data => {
         console.log("Tout les avocats:", JSON.stringify(data, null, 2));
         res.json(data);
@@ -15,25 +25,31 @@ exports.avocatsListe = async function (req, res) {
 }
 
 
-exports.avocatCreation = async function (req, res) {
+exports.lawyerCreation = async function (req, res) {
     let nom_avocat = req.body.nom_avocat;
     let coordonnees = req.body.coordonnees;
     let honoraires = req.body.honoraires;
     let photo = req.body.photo;
+    let is_admin = req.is_admin;
 
-    let avocat = Avocat.build({ nom_avocat: nom_avocat, coordonnees: coordonnees, honoraires: honoraires, photo: photo})
-    await avocat.save()
-    .then(data => {
-        console.log(avocat.toJSON());
-        res.json(data);
-    })
-    .catch(err => {
-        res.status(500).json({ message: err.message })
-    })
+    if(is_admin){
+        let avocat = Avocat.build({ nom_avocat: nom_avocat, coordonnees: coordonnees, honoraires: honoraires, photo: photo})
+        await avocat.save()
+        .then(data => {
+            console.log(avocat.toJSON());
+            res.json(data);
+        })
+        .catch(err => {
+            res.status(500).json({ message: err.message })
+        })
+    } else {
+        res.status(403).json({message: "Access denied"});
+    }
+
 }
 
 
-exports.avocatData = async function (req, res) {
+exports.lawyerData = async function (req, res) {
     let id_avocat = req.params.id_avocat;
     await Avocat.findAll({
         where: {id_avocat},
@@ -54,10 +70,12 @@ exports.avocatData = async function (req, res) {
 }
 
 
-exports.avocatDelete = async function (req, res) {
+exports.lawyerDelete = async function (req, res) {
     let id_avocat = req.params.id_avocat;
+    let is_admin = req.is_admin;
 
-    if (id_avocat) {
+
+    if (id_avocat && is_admin) {
         await Avocat.destroy({ where: 
             { id_avocat: id_avocat}
         })
@@ -69,18 +87,19 @@ exports.avocatDelete = async function (req, res) {
             res.status(500).json({ message: err.message })
         })
     }
-    else res.status(404).json({ message: "Lawyer doesn't exist"})
+    else res.status(404).json({ message: "Lawyer doesn't exist or access denied"})
 }
 
 
-exports.avocatModification = async function (req, res) {
+exports.lawyerModification = async function (req, res) {
     let id_avocat = req.body.id_avocat;
     let nom_avocat = req.body.nom_avocat;
     let coordonnees = req.body.coordonnees;
     let honoraires = req.body.honoraires;
     let photo = req.body.photo;
+    let is_admin = req.is_admin;
 
-    if (id_avocat) {
+    if (id_avocat && is_admin) {
         await Avocat.update({ 
             nom_avocat: nom_avocat,
             coordonnees: coordonnees,
@@ -96,15 +115,5 @@ exports.avocatModification = async function (req, res) {
             res.status(500).json({ message: err.message })
         })
     }
-    else res.status(404).json({ message: "Legal field doesn't exist"})
+    else res.status(404).json({ message: "Lawyer doesn't exist or access denied"})
 }
-
-// {
-//     "id_avocat": "<integer>",
-//     "nom_avocat": "<string>",
-//     "id_domaine": "<integer>",
-//     "nom_domaine": "<string>",
-//     "coordonnees": "<string>",
-//     "honoraires": "<string>",
-//     "photo": "<string>"
-//   }
